Move image files under `.~/get-started-page` to `.~/images`.
