+++ js/src/get-started-page/benefits-card/index.js
@@ -7,7 +7,7 @@ import { Card, CardBody, FlexItem } from '@wordpress/components';
 /**
  * Internal dependencies
  */
-import benefitsImageURL from './image.png';
+import benefitsImageURL from '.~/images/get-started/benefits.png';
 import Text from '.~/components/app-text';
 import './index.scss';
