+++ js/src/get-started-page/customer-quotes-card/index.js
@@ -8,7 +8,7 @@ import { Flex, FlexBlock, Card, CardHeader } from '@wordpress/components';
  * Internal dependencies
  */
 import './index.scss';
-import quoteImageURL from './img-quote.svg';
+import quoteImageURL from '.~/images/get-started/img-quote.svg';
 import Text from '.~/components/app-text';
 
 const Quote = ( { quote, name } ) => {
