+++ js/src/get-started-page/get-started-with-hero-card/index.js
@@ -14,7 +14,7 @@ import Text from '.~/components/app-text';
 import AppDocumentationLink from '.~/components/app-documentation-link';
 import { getSetupMCUrl } from '.~/utils/urls';
 import './index.scss';
-import heroUrl from './hero.png';
+import heroUrl from '.~/images/get-started/hero.png';
 
 /**
  * @fires gla_setup_mc with `{ triggered_by: 'start-onboarding-button', action: 'go-to-onboarding', context: 'get-started-with-hero' }`.
