+++ js/src/images/get-started/hero.png
